requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
langchain
pypdf
tiktoken
slowapi
python-multipart
mlflow
